add setweather method
