+++ src/scenes/levelScene.js
@@ -1,5 +1,6 @@
 import { gameState } from '../utils/gameState';
 import { config } from '../utils/config';
+import { setWeather } from '../utils/weather';
 
 export class Level extends Phaser.Scene {
     constructor(key) {
@@ -132,9 +133,8 @@ export class Level extends Phaser.Scene {
       for (const [xIndex, yIndex] of this.heights.entries()) {
         this.createPlatform(xIndex, yIndex);
       }
-  
-      // Create the door at the end of the level
-      gameState.goal = this.physics.add.sprite(gameState.width - 40, 100, 'door');
+ 
+      gameState.goal = this.physics.add.sprite(gameState.width - 20, 120, 'door').setScale(.4);
   
       this.physics.add.overlap(gameState.player, gameState.goal, function() {
         this.cameras.main.fade(800, 0, 0, 0, false, function(camera, progress) {
@@ -145,7 +145,7 @@ export class Level extends Phaser.Scene {
         });
       }, null, this);
   
-      this.setWeather(this.weather);
+      setWeather(this.weather, gameState);
     }
   
     update() {
@@ -196,56 +196,5 @@ export class Level extends Phaser.Scene {
         star.setScrollFactor(Math.random() * .1);
         gameState.stars.push(star)
       }
-    }
-  
-    setWeather(weather) {
-      const weathers = {
-  
-        'morning': {
-          'color': 0xecdccc,
-          'snow':  1,
-          'wind':  20,
-          'bgColor': 0xF8c3aC,
-        },
-  
-        'afternoon': {
-          'color': 0xffffff,
-          'snow':  1,
-          'wind': 80,
-          'bgColor': 0x0571FF,
-        },
-  
-        'twilight': {
-          'color': 0xccaacc,
-          'bgColor': 0x18235C,
-          'snow':  10,
-          'wind': 200,
-        },
-  
-        'night': {
-          'color': 0x555555,
-          'bgColor': 0x000000,
-          'snow':  0,
-          'wind': 0,
-        },
-      }
-      let { color, bgColor, snow, wind } = weathers[weather];
-      gameState.bg1.setTint(color);
-      gameState.bg2.setTint(color);
-      gameState.bg3.setTint(color);
-      gameState.bgColor.fillColor = bgColor;
-      gameState.emitter.setQuantity(snow);
-      gameState.emitter.setSpeedX(-wind);
-      gameState.player.setTint(color);
-      for (let platform of gameState.platforms.getChildren()) {
-        platform.setTint(color);
-      }
-      if (weather === 'night') {
-        gameState.stars.forEach(star => star.setVisible(true));
-      } else {
-        gameState.stars.forEach(star => star.setVisible(false));
-      }
-  
-      return
     }
   }
